* Move debug logs menu to the Application plugin. * How to trigger Editor.load in Application container?
